(mod/components)moved all sidebar components to its own dir
